+++ lib/Dialect/Torch/Transforms/AbstractInterpLibrary.cpp
@@ -8608,47 +8608,96 @@ StringRef mlir::torch::Torch::getAbstractInterpLibrary() {
 "    return %5 : !torch.int\n"
 "  }\n"
 "  func.func @__torch__.torch_mlir.dialects.torch.importer.jit_ir.build_tools.library_generator.get_priority_of_dtype(%arg0: !torch.int) -> !torch.int {\n"
-"    %true = torch.constant.bool true\n"
-"    %int12 = torch.constant.int 12\n"
 "    %none = torch.constant.none\n"
 "    %str = torch.constant.str \"AssertionError: Cannot determine priority of dtype\"\n"
-"    %int9223372036854775807 = torch.constant.int 9223372036854775807\n"
-"    %int10 = torch.constant.int 10\n"
-"    %int9 = torch.constant.int 9\n"
-"    %int7 = torch.constant.int 7\n"
-"    %int6 = torch.constant.int 6\n"
-"    %int5 = torch.constant.int 5\n"
 "    %int15 = torch.constant.int 15\n"
-"    %int4 = torch.constant.int 4\n"
-"    %int3 = torch.constant.int 3\n"
-"    %int2 = torch.constant.int 2\n"
-"    %int1 = torch.constant.int 1\n"
 "    %int0 = torch.constant.int 0\n"
+"    %int1 = torch.constant.int 1\n"
+"    %int2 = torch.constant.int 2\n"
+"    %int3 = torch.constant.int 3\n"
+"    %int4 = torch.constant.int 4\n"
+"    %int5 = torch.constant.int 5\n"
+"    %int6 = torch.constant.int 6\n"
+"    %int7 = torch.constant.int 7\n"
+"    %int8 = torch.constant.int 8\n"
+"    %int9 = torch.constant.int 9\n"
+"    %int10 = torch.constant.int 10\n"
 "    %int11 = torch.constant.int 11\n"
-"    %false = torch.constant.bool false\n"
 "    %0 = torch.prim.Uninitialized : !torch.int\n"
-"    %1 = torch.prim.ListConstruct %int11, %int0, %int1, %int2, %int3, %int4, %int15, %int5, %int6, %int7, %int9, %int10 : (!torch.int, !torch.int, !torch.int, !torch.int, !torch.int, !torch.int, !torch.int, !torch.int, !torch.int, !torch.int, !torch.int, !torch.int) -> !torch.list<int>\n"
-"    %2:3 = torch.prim.Loop %int9223372036854775807, %true, init(%false, %0, %int0) {\n"
-"    ^bb0(%arg1: !torch.int, %arg2: !torch.bool, %arg3: !torch.int, %arg4: !torch.int):\n"
-"      %4 = torch.aten.__getitem__.t %1, %arg4 : !torch.list<int>, !torch.int -> !torch.int\n"
-"      %5 = torch.aten.eq.int %4, %arg0 : !torch.int, !torch.int -> !torch.bool\n"
-"      %6:3 = torch.prim.If %5 -> (!torch.bool, !torch.bool, !torch.int) {\n"
-"        torch.prim.If.yield %false, %true, %arg4 : !torch.bool, !torch.bool, !torch.int\n"
-"      } else {\n"
-"        torch.prim.If.yield %true, %false, %0 : !torch.bool, !torch.bool, !torch.int\n"
-"      }\n"
-"      %7 = torch.aten.add.int %arg4, %int1 : !torch.int, !torch.int -> !torch.int\n"
-"      %8 = torch.aten.lt.int %7, %int12 : !torch.int, !torch.int -> !torch.bool\n"
-"      %9 = torch.aten.__and__.bool %8, %6#0 : !torch.bool, !torch.bool -> !torch.bool\n"
-"      torch.prim.Loop.condition %9, iter(%6#1, %6#2, %7 : !torch.bool, !torch.int, !torch.int)\n"
-"    } : (!torch.int, !torch.bool, !torch.bool, !torch.int, !torch.int) -> (!torch.bool, !torch.int, !torch.int)\n"
-"    %3 = torch.prim.If %2#0 -> (!torch.int) {\n"
-"      torch.prim.If.yield %2#1 : !torch.int\n"
+"    %1 = torch.aten.eq.int %arg0, %int11 : !torch.int, !torch.int -> !torch.bool\n"
+"    %2 = torch.prim.If %1 -> (!torch.int) {\n"
+"      torch.prim.If.yield %int0 : !torch.int\n"
 "    } else {\n"
-"      torch.prim.RaiseException %str, %none : !torch.str, !torch.none\n"
-"      torch.prim.If.yield %0 : !torch.int\n"
+"      %3 = torch.aten.eq.int %arg0, %int0 : !torch.int, !torch.int -> !torch.bool\n"
+"      %4 = torch.prim.If %3 -> (!torch.int) {\n"
+"        torch.prim.If.yield %int1 : !torch.int\n"
+"      } else {\n"
+"        %5 = torch.aten.eq.int %arg0, %int1 : !torch.int, !torch.int -> !torch.bool\n"
+"        %6 = torch.prim.If %5 -> (!torch.int) {\n"
+"          torch.prim.If.yield %int2 : !torch.int\n"
+"        } else {\n"
+"          %7 = torch.aten.eq.int %arg0, %int2 : !torch.int, !torch.int -> !torch.bool\n"
+"          %8 = torch.prim.If %7 -> (!torch.int) {\n"
+"            torch.prim.If.yield %int3 : !torch.int\n"
+"          } else {\n"
+"            %9 = torch.aten.eq.int %arg0, %int3 : !torch.int, !torch.int -> !torch.bool\n"
+"            %10 = torch.prim.If %9 -> (!torch.int) {\n"
+"              torch.prim.If.yield %int4 : !torch.int\n"
+"            } else {\n"
+"              %11 = torch.aten.eq.int %arg0, %int4 : !torch.int, !torch.int -> !torch.bool\n"
+"              %12 = torch.prim.If %11 -> (!torch.int) {\n"
+"                torch.prim.If.yield %int5 : !torch.int\n"
+"              } else {\n"
+"                %13 = torch.aten.eq.int %arg0, %int15 : !torch.int, !torch.int -> !torch.bool\n"
+"                %14 = torch.prim.If %13 -> (!torch.int) {\n"
+"                  torch.prim.If.yield %int6 : !torch.int\n"
+"                } else {\n"
+"                  %15 = torch.aten.eq.int %arg0, %int5 : !torch.int, !torch.int -> !torch.bool\n"
+"                  %16 = torch.prim.If %15 -> (!torch.int) {\n"
+"                    torch.prim.If.yield %int7 : !torch.int\n"
+"                  } else {\n"
+"                    %17 = torch.aten.eq.int %arg0, %int6 : !torch.int, !torch.int -> !torch.bool\n"
+"                    %18 = torch.prim.If %17 -> (!torch.int) {\n"
+"                      torch.prim.If.yield %int8 : !torch.int\n"
+"                    } else {\n"
+"                      %19 = torch.aten.eq.int %arg0, %int7 : !torch.int, !torch.int -> !torch.bool\n"
+"                      %20 = torch.prim.If %19 -> (!torch.int) {\n"
+"                        torch.prim.If.yield %int9 : !torch.int\n"
+"                      } else {\n"
+"                        %21 = torch.aten.eq.int %arg0, %int9 : !torch.int, !torch.int -> !torch.bool\n"
+"                        %22 = torch.prim.If %21 -> (!torch.int) {\n"
+"                          torch.prim.If.yield %int10 : !torch.int\n"
+"                        } else {\n"
+"                          %23 = torch.aten.eq.int %arg0, %int10 : !torch.int, !torch.int -> !torch.bool\n"
+"                          %24 = torch.prim.If %23 -> (!torch.int) {\n"
+"                            torch.prim.If.yield %int11 : !torch.int\n"
+"                          } else {\n"
+"                            torch.prim.RaiseException %str, %none : !torch.str, !torch.none\n"
+"                            torch.prim.If.yield %0 : !torch.int\n"
+"                          }\n"
+"                          torch.prim.If.yield %24 : !torch.int\n"
+"                        }\n"
+"                        torch.prim.If.yield %22 : !torch.int\n"
+"                      }\n"
+"                      torch.prim.If.yield %20 : !torch.int\n"
+"                    }\n"
+"                    torch.prim.If.yield %18 : !torch.int\n"
+"                  }\n"
+"                  torch.prim.If.yield %16 : !torch.int\n"
+"                }\n"
+"                torch.prim.If.yield %14 : !torch.int\n"
+"              }\n"
+"              torch.prim.If.yield %12 : !torch.int\n"
+"            }\n"
+"            torch.prim.If.yield %10 : !torch.int\n"
+"          }\n"
+"          torch.prim.If.yield %8 : !torch.int\n"
+"        }\n"
+"        torch.prim.If.yield %6 : !torch.int\n"
+"      }\n"
+"      torch.prim.If.yield %4 : !torch.int\n"
 "    }\n"
-"    return %3 : !torch.int\n"
+"    return %2 : !torch.int\n"
 "  }\n"
 "  func.func @\"__torch_mlir_dtype_fn.aten.div.Tensor\"(%arg0: !torch.tuple<int, int>, %arg1: !torch.tuple<int, int>) -> !torch.int {\n"
 "    %false = torch.constant.bool false\n"
